requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
streamlit>=1.22.0
opencv-python>=4.7.0
pillow>=9.5.0
tqdm>=4.65.0
plotly>=5.14.0
pandas>=2.0.0
albumentations>=1.3.0
seaborn>=0.12.0
openpyxl>=3.1.0
requests>=2.30.0
python-dotenv>=1.0.0
pyyaml>=6.0.0